4가지 방식 인증 구현/ test해봐야함
일단 DB연동해서 인증하는건 했는데 서버에서 쿼리문 돌렸을때 정확한 계산이 됐는지 의문임 내일 모여서 확인해볼것

+++ src/Real_VisitTogether_v3/app/src/main/java/authentication/Geodegree.java
@@ -4,7 +4,7 @@ import android.media.ExifInterface;
 
 public class Geodegree {
     private boolean valid = false;
-    private Float latitude, longitude;
+    private double latitude, longitude;
 
     public Geodegree(ExifInterface exif) {
         String attrLATITUDE = exif.getAttribute(ExifInterface.TAG_GPS_LATITUDE);
@@ -64,11 +64,11 @@ public class Geodegree {
         return ("위도 : " + String.valueOf(latitude) + "\n" + "경도 : " + String.valueOf(longitude));
     }
 
-    public Float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public Float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
